# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: gameManageKit API
  version: 1.2.0
  description: 多游戏、HTTP-only、MySQL-only 的账号门户服务。
servers:
  - url: https://portal.example.com
    description: Public
  - url: http://game-manage-kit.internal:2571
    description: Internal
...
paths:
...
  /v1/games/{gameId}/admin/accounts/{userId}/ban:
    parameters:
      - $ref: '#/components/parameters/GameId'
      - $ref: '#/components/parameters/UserId'
    post:
      operationId: banAccount
      tags: [Admin]
      security:
        - AdminSession: []
        - AdminSecret: []
          OperatorId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminAccountRequest'
      responses:
        '200':
          description: 封号结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/games/{gameId}/admin/accounts/{userId}/revoke:
    parameters:
      - $ref: '#/components/parameters/GameId'
      - $ref: '#/components/parameters/UserId'
    post:
      operationId: revokeAccount
      tags: [Admin]
      security:
        - AdminSession: []
        - AdminSecret: []
          OperatorId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminAccountRequest'
      responses:
        '200':
          description: 撤销该游戏账号的全部现有会话；账号保留且允许重新登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/games/{gameId}/admin/accounts/{userId}:
    parameters:
      - $ref: '#/components/parameters/GameId'
      - $ref: '#/components/parameters/UserId'
    get:
      operationId: getAdminAccount
      tags: [Admin]
      security:
        - AdminSession: []
        - AdminSecret: []
          OperatorId: []
      responses:
        '200':
          description: 仅返回管理所需的最小账号摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccountDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    UserBearer:
      type: http
      scheme: bearer
    ServiceId:
      type: apiKey
      in: header
      name: x-service-id
    ServiceSecret:
      type: apiKey
      in: header
      name: x-service-secret
    AdminSecret:
      type: apiKey
      in: header
      name: x-admin-secret
    OperatorId:
      type: apiKey
      in: header
      name: x-operator-id
    AdminSession:
      type: apiKey
      in: cookie
      name: __Host-gmk_admin_session
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/GameId'
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 32
        pattern: '^u_[0-9]+$'
...
  responses:
    BadRequest:
      description: 入参不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    AdminUnauthorized:
      description: 管理员凭证或会话无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 禁止访问；可能为账号封禁、游戏停用、调用方无游戏权限或区服停用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源、游戏或区服不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 操作冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: 请求过频
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Internal:
      description: 内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unavailable:
      description: 微信上游不可用，或游戏处于维护状态
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    GameId:
      type: string
      minLength: 2
      maxLength: 32
      pattern: '^[a-z][a-z0-9-]{1,31}$'
...
    ErrorCode:
      type: string
      enum:
        - INVALID_PAYLOAD
        - AUTH_REQUIRED
        - ACCOUNT_BANNED
        - NOT_FOUND
        - RATE_LIMITED
        - UPSTREAM_UNAVAILABLE
        - SERVICE_AUTH_REQUIRED
        - SERVICE_FORBIDDEN
        - OPERATION_CONFLICT
        - INTERNAL
        - GAME_NOT_FOUND
        - GAME_DISABLED
        - GAME_ACCESS_DENIED
        - SERVER_NOT_FOUND
        - SERVER_DISABLED
        - ADMIN_AUTH_REQUIRED
        - ORIGIN_FORBIDDEN
        - GAME_PROJECT_CONFLICT
        - GAME_SERVER_CONFLICT
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, requestId]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        requestId:
          type: string
...
    AdminAccountRequest:
      type: object
      additionalProperties: false
      required: [operationId, reason]
      properties:
        operationId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_.:-]+$'
        reason:
          type: string
          minLength: 1
          maxLength: 255
...
    AdminAccountDetailResponse:
      type: object
      additionalProperties: false
      required: [userId, status, lastLoginAt, activeSessionCount]
      properties:
        userId:
          type: string
          minLength: 1
          maxLength: 32
          pattern: '^u_[0-9]+$'
        status:
          type: string
          enum: [active, banned, deregistered]
        lastLoginAt:
          type:
            - string
            - 'null'
          format: date-time
        activeSessionCount:
          type: integer
          minimum: 0
    AdminAccountResponse:
      type: object
      additionalProperties: false
      required: [accountExists, status]
      properties:
        accountExists:
          type: boolean
        status:
          type: string
          enum: [banned, revoked, not_found]
...